amountRaised closes #21
