+++ package.json
@@ -39,6 +39,7 @@
     "react-dom": "^15.3.1",
     "react-router": "^2.7.0",
     "react-stormpath": "^1.3.1",
+    "stripe": "^4.11.0",
     "superagent": "^2.2.0"
   }
 }
